The file picker is working and I'm reordering the code to make it easier to follow
