+++ app/src/main/java/org/schabi/newpipe/SettingsActivity.java
@@ -4,20 +4,11 @@ import android.app.Activity;
 import android.content.ClipData;
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.content.res.Configuration;
-import android.media.AudioManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
-import android.os.Environment;
-import android.preference.CheckBoxPreference;
-import android.preference.EditTextPreference;
-import android.preference.ListPreference;
-import android.preference.Preference;
 import android.preference.PreferenceActivity;
-import android.preference.PreferenceFragment;
-import android.preference.PreferenceManager;
 import android.support.annotation.LayoutRes;
 import android.support.annotation.NonNull;
 import android.support.v7.app.ActionBar;
@@ -27,13 +18,11 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.nononsenseapps.filepicker.AbstractFilePickerFragment;
 import com.nononsenseapps.filepicker.FilePickerActivity;
 
 import java.util.ArrayList;
 
 import info.guardianproject.netcipher.proxy.OrbotHelper;
-import us.shandian.giga.util.Utility;
 
 /**
  * Created by Christian Schabesberger on 31.08.15.
@@ -57,8 +46,8 @@ import us.shandian.giga.util.Utility;
 
 public class SettingsActivity extends PreferenceActivity  {
 
-    private static final int REQUEST_INSTALL_ORBOT = 0x1234;
-    private AppCompatDelegate mDelegate;
+    public static final int REQUEST_INSTALL_ORBOT = 0x1234;
+    private AppCompatDelegate mDelegate = null;
 
     @Override
     protected void onCreate(Bundle savedInstanceBundle) {
@@ -77,117 +66,6 @@ public class SettingsActivity extends PreferenceActivity  {
 
     }
 
-    public static class SettingsFragment extends PreferenceFragment{
-        SharedPreferences.OnSharedPreferenceChangeListener prefListener;
-
-        // get keys
-        String DEFAULT_RESOLUTION_PREFERENCE;
-        String DEFAULT_AUDIO_FORMAT_PREFERENCE;
-        String SEARCH_LANGUAGE_PREFERENCE;
-        String DOWNLOAD_PATH_PREFERENCE;
-        String DOWNLOAD_PATH_AUDIO_PREFERENCE;
-        String USE_TOR_KEY;
-
-        private ListPreference defaultResolutionPreference;
-        private ListPreference defaultAudioFormatPreference;
-        private ListPreference searchLanguagePreference;
-        private Preference downloadPathPreference;
-        private EditTextPreference downloadPathAudioPreference;
-        private CheckBoxPreference useTorCheckBox;
-        private SharedPreferences defaultPreferences;
-
-
-        @Override
-        public void onCreate(Bundle savedInstanceState) {
-            super.onCreate(savedInstanceState);
-            addPreferencesFromResource(R.xml.settings);
-
-            final Activity activity = getActivity();
-
-            defaultPreferences = PreferenceManager.getDefaultSharedPreferences(activity);
-
-            // get keys
-            DEFAULT_RESOLUTION_PREFERENCE =getString(R.string.default_resolution_key);
-            DEFAULT_AUDIO_FORMAT_PREFERENCE =getString(R.string.default_audio_format_key);
-            SEARCH_LANGUAGE_PREFERENCE =getString(R.string.search_language_key);
-            DOWNLOAD_PATH_PREFERENCE = getString(R.string.download_path_key);
-            DOWNLOAD_PATH_AUDIO_PREFERENCE = getString(R.string.download_path_audio_key);
-            USE_TOR_KEY = getString(R.string.use_tor_key);
-
-            // get pref objects
-            defaultResolutionPreference =
-                    (ListPreference) findPreference(DEFAULT_RESOLUTION_PREFERENCE);
-            defaultAudioFormatPreference =
-                    (ListPreference) findPreference(DEFAULT_AUDIO_FORMAT_PREFERENCE);
-            searchLanguagePreference =
-                    (ListPreference) findPreference(SEARCH_LANGUAGE_PREFERENCE);
-            downloadPathPreference = findPreference(DOWNLOAD_PATH_PREFERENCE);
-            downloadPathAudioPreference =
-                    (EditTextPreference) findPreference(DOWNLOAD_PATH_AUDIO_PREFERENCE);
-            useTorCheckBox = (CheckBoxPreference) findPreference(USE_TOR_KEY);
-
-            prefListener = new SharedPreferences.OnSharedPreferenceChangeListener() {
-                @Override
-                public void onSharedPreferenceChanged(SharedPreferences sharedPreferences,
-                                                      String key) {
-                    Activity a = getActivity();
-                    if(a != null) {
-                        updateSummary();
-
-                        if (defaultPreferences.getBoolean(USE_TOR_KEY, false)) {
-                            if (OrbotHelper.isOrbotInstalled(a)) {
-                                App.configureTor(true);
-                                OrbotHelper.requestStartTor(a);
-                            } else {
-                                Intent intent = OrbotHelper.getOrbotInstallIntent(a);
-                                a.startActivityForResult(intent, REQUEST_INSTALL_ORBOT);
-                            }
-                        } else {
-                            App.configureTor(false);
-                        }
-                    }
-                }
-            };
-            defaultPreferences.registerOnSharedPreferenceChangeListener(prefListener);
-
-            downloadPathPreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener(){
-                @Override
-                public boolean onPreferenceClick(Preference preference) {
-                    Intent i = new Intent(getActivity(), FilePickerActivity.class);
-
-                    i.putExtra(FilePickerActivity.EXTRA_ALLOW_MULTIPLE, false);
-                    i.putExtra(FilePickerActivity.EXTRA_ALLOW_CREATE_DIR, true);
-                    i.putExtra(FilePickerActivity.EXTRA_MODE, FilePickerActivity.MODE_DIR);
-                    startActivityForResult(i, 233);
-
-                    return true;
-                }
-            });
-
-
-            updateSummary();
-        }
-
-        // This is used to show the status of some preference in the description
-        private void updateSummary() {
-            defaultResolutionPreference.setSummary(
-                    defaultPreferences.getString(DEFAULT_RESOLUTION_PREFERENCE,
-                            getString(R.string.default_resolution_value)));
-            defaultAudioFormatPreference.setSummary(
-                    defaultPreferences.getString(DEFAULT_AUDIO_FORMAT_PREFERENCE,
-                            getString(R.string.default_audio_format_value)));
-            searchLanguagePreference.setSummary(
-                    defaultPreferences.getString(SEARCH_LANGUAGE_PREFERENCE,
-                            getString(R.string.default_language_value)));
-            downloadPathPreference.setSummary(
-                    defaultPreferences.getString(DOWNLOAD_PATH_PREFERENCE,
-                            getString(R.string.download_path_summary)));
-            downloadPathAudioPreference.setSummary(
-                    defaultPreferences.getString(DOWNLOAD_PATH_AUDIO_PREFERENCE,
-                            getString(R.string.download_path_audio_summary)));
-        }
-    }
-
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
 
@@ -200,7 +78,6 @@ public class SettingsActivity extends PreferenceActivity  {
                     if (clip != null) {
                         for (int i = 0; i < clip.getItemCount(); i++) {
                             Uri uri = clip.getItemAt(i).getUri();
-                            // Do something with the URI
                         }
                     }
                     // For Ice Cream Sandwich
@@ -211,25 +88,22 @@ public class SettingsActivity extends PreferenceActivity  {
                     if (paths != null) {
                         for (String path: paths) {
                             Uri uri = Uri.parse(path);
-                            // Do something with the URI
                         }
                     }
                 }
 
             } else {
                 Uri uri = data.getData();
-                // Do something with the URI
             }
-
-
+        }
+        else if(requestCode == REQUEST_INSTALL_ORBOT)
+        {
+            // try to start tor regardless of resultCode since clicking back after
+            // installing the app does not necessarily return RESULT_OK
+            App.configureTor(requestCode == REQUEST_INSTALL_ORBOT
+                    && OrbotHelper.requestStartTor(this));
 
         }
-
-        // try to start tor regardless of resultCode since clicking back after
-        // installing the app does not necessarily return RESULT_OK
-        App.configureTor(requestCode == REQUEST_INSTALL_ORBOT
-                && OrbotHelper.requestStartTor(this));
-
         super.onActivityResult(requestCode, resultCode, data);
     }
